feat(task,summon-core): extract @canonical/task, restructure summon as @canonical/summon-core (v0.1-P1+P2)
New package @canonical/task at packages/runtime/task/ (src/lib/ convention):
- Task monad, Effect union, TaskError, all monad operations, combinators,
  primitives, effect constructors, interpreter (domain-agnostic), dry-run,
  test utilities. 524 tests.

Restructured @canonical/summon → @canonical/summon-core:
- Deleted duplicated effect framework files (types, task, effect, primitives,
  combinators, interpreter, dry-run) and their tests
- Internal modules (template, cli, components) import from @canonical/task
- Extracted stamp logic from interpreter into stamp.ts as pure composition
  (applyStamp + runTaskWithStamp wrapping runTask)
- Package renamed to @canonical/summon-core

Downstream packages (summon-component, summon-package, summon-monorepo):
- Effect framework imports from @canonical/task
- Generator types from @canonical/summon-core

Decisions: PA.01, PA.04, PA.05, PA.06, PA.14, PA.15
