+++ package.json
@@ -20,6 +20,7 @@
     "configs/*",
     "packages/*",
     "packages/react/*",
+    "packages/runtime/*",
     "packages/storybook/*",
     "packages/summon/*",
     "packages/styles/*",
