+++ packages/runtime/task/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@canonical/summon-core",
-  "description": "Code generation framework: generator definitions, templates, stamps. Built on @canonical/task.",
+  "name": "@canonical/task",
+  "description": "A monadic effect framework for composable, testable, dry-runnable CLI operations.",
   "version": "0.18.0",
   "type": "module",
   "module": "src/index.ts",
@@ -11,12 +11,8 @@
       "import": "./src/index.ts"
     }
   },
-  "bin": {
-    "summon": "src/cli.tsx"
-  },
   "files": [
-    "src",
-    "generators"
+    "src"
   ],
   "author": {
     "email": "webteam@canonical.com",
@@ -34,12 +30,11 @@
   "scripts": {
     "build": "echo 'No build needed - runs directly from TypeScript'",
     "build:all": "bun run build",
-    "check": "bun run check:biome && bun run check:ts && bun run check:webarchitect",
+    "check": "bun run check:biome && bun run check:ts",
     "check:fix": "bun run check:biome:fix && bun run check:ts",
     "check:biome": "biome check",
     "check:biome:fix": "biome check --write",
     "check:ts": "tsc --noEmit",
-    "check:webarchitect": "webarchitect tool-ts",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage"
@@ -48,26 +43,9 @@
     "@biomejs/biome": "2.4.6",
     "@canonical/biome-config": "^0.18.0",
     "@canonical/typescript-config": "^0.18.0",
-    "@canonical/webarchitect": "^0.18.0",
     "@types/bun": "^1.3.10",
-    "@types/ejs": "^3.1.5",
     "@types/node": "^24.12.0",
-    "@types/omelette": "^0.4.5",
-    "@types/react": "^19.2.14",
-    "ink-testing-library": "^4.0.0",
     "typescript": "^5.9.3",
     "vitest": "^4.0.18"
-  },
-  "dependencies": {
-    "@canonical/task": "^0.18.0",
-    "chalk": "^5.6.2",
-    "commander": "^14.0.3",
-    "ejs": "^4.0.1",
-    "ink": "^6.8.0",
-    "ink-select-input": "^6.2.0",
-    "ink-spinner": "^5.0.0",
-    "ink-text-input": "^6.0.0",
-    "omelette": "^0.4.17",
-    "react": "^19.2.4"
   }
 }
