+++ packages/runtime/task/src/__tests__/primitives.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { dryRun, dryRunWith } from "../dry-run.js";
+import { dryRun, dryRunWith } from "../lib/dry-run.js";
 import {
   copyDirectory,
   copyFile,
@@ -28,8 +28,8 @@ import {
   warn,
   withContext,
   writeFile,
-} from "../primitives.js";
-import { flatMap } from "../task.js";
+} from "../lib/primitives.js";
+import { flatMap } from "../lib/task.js";
 
 // =============================================================================
 // File System Primitives
