+++ packages/runtime/task/src/lib/interpreter.ts
@@ -10,114 +10,6 @@ import * as path from "node:path";
 import type { Effect, ExecResult, Task, TaskError } from "./types.js";
 
 // =============================================================================
-// Stamp Options
-// =============================================================================
-
-export interface StampConfig {
-  /** Generator name (e.g., "@canonical/summon-package") */
-  generator: string;
-  /** Generator version (e.g., "0.1.0") */
-  version: string;
-}
-
-// Comment styles for stamp generation
-const COMMENT_STYLES: Record<
-  string,
-  {
-    single?: string;
-    blockStart?: string;
-    blockEnd?: string;
-    preferBlock?: boolean;
-  }
-> = {
-  // JavaScript/TypeScript
-  ".ts": { single: "//" },
-  ".tsx": { single: "//" },
-  ".js": { single: "//" },
-  ".jsx": { single: "//" },
-  ".mjs": { single: "//" },
-  ".cjs": { single: "//" },
-  // CSS
-  ".css": { blockStart: "/*", blockEnd: "*/", preferBlock: true },
-  ".scss": { single: "//" },
-  ".sass": { single: "//" },
-  ".less": { single: "//" },
-  // HTML/XML
-  ".html": { blockStart: "<!--", blockEnd: "-->" },
-  ".htm": { blockStart: "<!--", blockEnd: "-->" },
-  ".xml": { blockStart: "<!--", blockEnd: "-->" },
-  ".svg": { blockStart: "<!--", blockEnd: "-->" },
-  ".vue": { blockStart: "<!--", blockEnd: "-->" },
-  ".svelte": { blockStart: "<!--", blockEnd: "-->" },
-  // Config (YAML/TOML support comments, JSON does not)
-  ".yaml": { single: "#" },
-  ".yml": { single: "#" },
-  ".toml": { single: "#" },
-  // Shell/scripting
-  ".sh": { single: "#" },
-  ".bash": { single: "#" },
-  ".zsh": { single: "#" },
-  ".py": { single: "#" },
-  ".rb": { single: "#" },
-  // Other languages
-  ".go": { single: "//" },
-  ".rs": { single: "//" },
-  ".java": { single: "//" },
-  ".kt": { single: "//" },
-  ".swift": { single: "//" },
-  ".c": { single: "//" },
-  ".cpp": { single: "//" },
-  ".h": { single: "//" },
-  ".php": { single: "//" },
-  // Documentation
-  ".md": { blockStart: "<!--", blockEnd: "-->" },
-  ".mdx": { blockStart: "{/*", blockEnd: "*/}" },
-  // SQL
-  ".sql": { single: "--" },
-};
-
-/**
- * Apply a stamp to file content based on file extension.
- * Returns the original content if the file type doesn't support comments.
- */
-const applyStamp = (
-  filePath: string,
-  content: string,
-  stamp: StampConfig,
-): string => {
-  const ext = path.extname(filePath).toLowerCase();
-  const style = COMMENT_STYLES[ext];
-
-  if (!style) {
-    return content;
-  }
-
-  const stampText = `Generated by ${stamp.generator} v${stamp.version}`;
-  let stampLine: string;
-
-  if (style.preferBlock || (!style.single && style.blockStart)) {
-    if (!style.blockStart || !style.blockEnd) return content;
-    stampLine = `${style.blockStart} ${stampText} ${style.blockEnd}`;
-  } else if (style.single) {
-    stampLine = `${style.single} ${stampText}`;
-  } else {
-    return content;
-  }
-
-  // Handle shebang - place stamp after it
-  if (content.startsWith("#!")) {
-    const firstNewline = content.indexOf("\n");
-    if (firstNewline !== -1) {
-      const shebang = content.slice(0, firstNewline + 1);
-      const rest = content.slice(firstNewline + 1);
-      return `${shebang}${stampLine}\n${rest}`;
-    }
-  }
-
-  return `${stampLine}\n${content}`;
-};
-
-// =============================================================================
 // Task Execution Error
 // =============================================================================
 
@@ -150,7 +42,6 @@ export const executeEffect = async (
   context: Map<string, unknown>,
   promptHandler?: (question: Effect & { _tag: "Prompt" }) => Promise<unknown>,
   onLog?: (level: "debug" | "info" | "warn" | "error", message: string) => void,
-  stamp?: StampConfig,
 ): Promise<unknown> => {
   switch (effect._tag) {
     case "ReadFile": {
@@ -160,11 +51,7 @@ export const executeEffect = async (
     case "WriteFile": {
       const dir = path.dirname(effect.path);
       await fs.mkdir(dir, { recursive: true });
-      // Apply stamp if configured
-      const content = stamp
-        ? applyStamp(effect.path, effect.content, stamp)
-        : effect.content;
-      await fs.writeFile(effect.path, content, "utf-8");
+      await fs.writeFile(effect.path, effect.content, "utf-8");
       return undefined;
     }
 
@@ -367,8 +254,6 @@ export interface RunTaskOptions {
   onEffectComplete?: (effect: Effect, duration: number) => void;
   /** Handler for log effects. If provided, log output goes here instead of console */
   onLog?: (level: "debug" | "info" | "warn" | "error", message: string) => void;
-  /** Stamp configuration for generated files. If provided, all written files get a stamp comment. */
-  stamp?: StampConfig;
 }
 
 // =============================================================================
@@ -388,7 +273,6 @@ export const runTask = async <A>(
     onEffectStart,
     onEffectComplete,
     onLog,
-    stamp,
   } = options;
 
   const runInternal = async <B>(t: Task<B>): Promise<B> => {
@@ -440,7 +324,6 @@ export const runTask = async <A>(
           context,
           promptHandler,
           onLog,
-          stamp,
         );
 
         const duration = performance.now() - startTime;
@@ -455,7 +338,7 @@ export const runTask = async <A>(
 };
 
 /**
- * Run a task with a fresh context (simple API for backward compatibility).
+ * Run a task with a fresh context (simple API).
  */
 export const run = <A>(
   task: Task<A>,
