+++ packages/summon/core/src/stamp.ts
@@ -1,0 +1,165 @@
+/**
+ * Stamp — Generated File Stamp Logic
+ *
+ * Stamps are a generation concern: a comment prepended to generated files
+ * identifying the generator that produced them. The base interpreter
+ * (@canonical/task) is domain-agnostic and does not know about stamps.
+ *
+ * Stamps are applied via composition: `runTaskWithStamp` wraps `runTask`
+ * from @canonical/task, intercepting WriteFile effects to prepend stamps.
+ */
+
+import * as path from "node:path";
+import {
+  type Effect,
+  type RunTaskOptions,
+  runTask,
+  type Task,
+} from "@canonical/task";
+
+// =============================================================================
+// Stamp Config
+// =============================================================================
+
+export interface StampConfig {
+  /** Generator name (e.g., "@canonical/summon-package") */
+  generator: string;
+  /** Generator version (e.g., "0.1.0") */
+  version: string;
+}
+
+// =============================================================================
+// Comment Styles
+// =============================================================================
+
+const COMMENT_STYLES: Record<
+  string,
+  {
+    single?: string;
+    blockStart?: string;
+    blockEnd?: string;
+    preferBlock?: boolean;
+  }
+> = {
+  ".ts": { single: "//" },
+  ".tsx": { single: "//" },
+  ".js": { single: "//" },
+  ".jsx": { single: "//" },
+  ".mjs": { single: "//" },
+  ".cjs": { single: "//" },
+  ".css": { blockStart: "/*", blockEnd: "*/", preferBlock: true },
+  ".scss": { single: "//" },
+  ".sass": { single: "//" },
+  ".less": { single: "//" },
+  ".html": { blockStart: "<!--", blockEnd: "-->" },
+  ".htm": { blockStart: "<!--", blockEnd: "-->" },
+  ".xml": { blockStart: "<!--", blockEnd: "-->" },
+  ".svg": { blockStart: "<!--", blockEnd: "-->" },
+  ".vue": { blockStart: "<!--", blockEnd: "-->" },
+  ".svelte": { blockStart: "<!--", blockEnd: "-->" },
+  ".yaml": { single: "#" },
+  ".yml": { single: "#" },
+  ".toml": { single: "#" },
+  ".sh": { single: "#" },
+  ".bash": { single: "#" },
+  ".zsh": { single: "#" },
+  ".py": { single: "#" },
+  ".rb": { single: "#" },
+  ".go": { single: "//" },
+  ".rs": { single: "//" },
+  ".java": { single: "//" },
+  ".kt": { single: "//" },
+  ".swift": { single: "//" },
+  ".c": { single: "//" },
+  ".cpp": { single: "//" },
+  ".h": { single: "//" },
+  ".php": { single: "//" },
+  ".md": { blockStart: "<!--", blockEnd: "-->" },
+  ".mdx": { blockStart: "{/*", blockEnd: "*/}" },
+  ".sql": { single: "--" },
+};
+
+// =============================================================================
+// Stamp Application (Pure Function)
+// =============================================================================
+
+/**
+ * Apply a stamp to file content based on file extension.
+ * Returns the original content if the file type doesn't support comments.
+ */
+export const applyStamp = (
+  filePath: string,
+  content: string,
+  stamp: StampConfig,
+): string => {
+  const ext = path.extname(filePath).toLowerCase();
+  const style = COMMENT_STYLES[ext];
+
+  if (!style) {
+    return content;
+  }
+
+  const stampText = `Generated by ${stamp.generator} v${stamp.version}`;
+  let stampLine: string;
+
+  if (style.preferBlock || (!style.single && style.blockStart)) {
+    if (!style.blockStart || !style.blockEnd) return content;
+    stampLine = `${style.blockStart} ${stampText} ${style.blockEnd}`;
+  } else if (style.single) {
+    stampLine = `${style.single} ${stampText}`;
+  } else {
+    return content;
+  }
+
+  // Handle shebang - place stamp after it
+  if (content.startsWith("#!")) {
+    const firstNewline = content.indexOf("\n");
+    if (firstNewline !== -1) {
+      const shebang = content.slice(0, firstNewline + 1);
+      const rest = content.slice(firstNewline + 1);
+      return `${shebang}${stampLine}\n${rest}`;
+    }
+  }
+
+  return `${stampLine}\n${content}`;
+};
+
+// =============================================================================
+// Stamped Task Runner
+// =============================================================================
+
+export interface RunTaskWithStampOptions extends RunTaskOptions {
+  /** Stamp configuration for generated files. If provided, all written files get a stamp comment. */
+  stamp?: StampConfig;
+}
+
+/**
+ * Run a task with stamp support. Wraps `runTask` from @canonical/task,
+ * intercepting WriteFile effects to prepend generated-file stamps.
+ */
+export const runTaskWithStamp = async <A>(
+  task: Task<A>,
+  options: RunTaskWithStampOptions = {},
+): Promise<A> => {
+  const { stamp, onEffectStart: userOnEffectStart, ...restOptions } = options;
+
+  if (!stamp) {
+    return runTask(task, options);
+  }
+
+  // Wrap onEffectStart to mutate WriteFile content in-place before execution.
+  // This is the composition point: stamps are applied as a pre-processing step
+  // on the effect data, not baked into the interpreter.
+  const onEffectStart = (effect: Effect) => {
+    if (effect._tag === "WriteFile") {
+      (effect as { content: string }).content = applyStamp(
+        effect.path,
+        effect.content,
+        stamp,
+      );
+    }
+    userOnEffectStart?.(effect);
+  };
+
+  return runTask(task, { ...restOptions, onEffectStart });
+};
