+++ packages/summon/core/src/template.ts
@@ -9,11 +9,16 @@
  */
 
 import * as path from "node:path";
+import {
+  glob,
+  mkdir,
+  readFile,
+  sequence_,
+  type Task,
+  task,
+  writeFile,
+} from "@canonical/task";
 import ejs from "ejs";
-import { sequence_ } from "./combinators.js";
-import { glob, mkdir, readFile, writeFile } from "./primitives.js";
-import { task } from "./task.js";
-import type { Task } from "./types.js";
 
 // =============================================================================
 // Templating Engine Interface
