+++ packages/summon/package/src/package/index.ts
@@ -7,16 +7,11 @@
 import * as path from "node:path";
 import { fileURLToPath } from "node:url";
 import {
-  exec,
-  flatMap,
   type GeneratorDefinition,
-  info,
-  mkdir,
   type PromptDefinition,
-  sequence_,
   template,
-  when,
-} from "@canonical/summon";
+} from "@canonical/summon-core";
+import { exec, flatMap, info, mkdir, sequence_, when } from "@canonical/task";
 
 import {
   createTemplateContext,
